[NET-687][NET-817] Fix object log table and component placement
