+++ app/src/app/framework/object/components/object-log-list/object-log-change-view/object-log-change-view.component.html
