+++ app/src/app/framework/object/components/object-log-list/object-log-change-view/object-log-change-view.component.ts
@@ -1,6 +1,6 @@
 /*
 * DATAGERRY - OpenSource Enterprise CMDB
-* Copyright (C) 2019 NETHINKS GmbH
+* Copyright (C) 2019 - 2021 NETHINKS GmbH
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as
@@ -13,11 +13,11 @@
 * GNU Affero General Public License for more details.
 
 * You should have received a copy of the GNU Affero General Public License
-* along with this program.  If not, see <https://www.gnu.org/licenses/>.
+* along with this program. If not, see <https://www.gnu.org/licenses/>.
 */
 
 import { Component, Input } from '@angular/core';
-import { LogMode } from '../../../../../modes.enum';
+import { LogMode } from '../../../../modes.enum';
 
 @Component({
   selector: 'cmdb-object-log-change-view',
@@ -30,8 +30,5 @@ export class ObjectLogChangeViewComponent {
   @Input() action: LogMode;
   @Input() changes: any[];
 
-  public constructor() {
-
-  }
 
 }
