+++ app/src/app/framework/object/components/object-log-list/object-log-list.component.html
