+++ app/src/app/framework/object/components/object-log-list/object-log-user/object-log-user.component.html
